refactor: rename files, where definitions changed
