+++ packages/tool-server/src/tools/interactions/gesture-swipe.ts
