+++ packages/tool-server/src/utils/setup-registry.ts
@@ -11,9 +11,9 @@ import { restartAppTool } from "../tools/simulator/restart-app";
 import { reinstallAppTool } from "../tools/simulator/reinstall-app";
 import { openUrlTool } from "../tools/simulator/open-url";
 import { screenshotTool } from "../tools/interactions/screenshot";
-import { gestureTapTool } from "../tools/interactions/tap";
-import { gestureSwipeTool } from "../tools/interactions/swipe";
-import { gestureCustomTool } from "../tools/interactions/gesture";
+import { gestureTapTool } from "../tools/interactions/gesture-tap";
+import { gestureSwipeTool } from "../tools/interactions/gesture-swipe";
+import { gestureCustomTool } from "../tools/interactions/gesture-custom";
 import { gesturePinchTool } from "../tools/interactions/gesture-pinch";
 import { gestureRotateTool } from "../tools/interactions/gesture-rotate";
 import { buttonTool } from "../tools/interactions/button";
